requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
babel==2.17.0
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dnspython==2.8.0
docopt==0.6.2
docxcompose==1.4.0
docxtpl==0.16.7
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.122.0
fonttools==4.60.1
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.2
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
motor==3.7.1
mysql-connector-python==8.2.0
num2words==0.5.13
numpy==1.26.4
openpyxl==3.1.2
packaging==25.0
#pandas==2.1.4
passlib==1.7.4
pillow==12.0.0
protobuf==4.21.12
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.15.4
pyparsing==3.2.5
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.1
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1


# Dependencias del Backend - Sistema de Autenticación y Roles

# Framework web
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Base de datos
motor>=3.3.2  # Async driver para MongoDB
pymongo>=4.6.0  # Driver síncrono adicional (para seeders)

# Autenticación y seguridad
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.1

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic[email]>=2.5.0

# HTTP requests (para API externa)
requests>=2.31.0

# CORS
python-multipart>=0.0.6
